+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Remove the `spool` command from Relay. ([#4423](https://github.com/getsentry/relay/pull/4423))
 - Bump `sentry-native` to `0.7.17` and remove cross compilation in CI. ([#4427](https://github.com/getsentry/relay/pull/4427))
 - Remove `form_data` envelope items from standalone envelopes. ([#4428](https://github.com/getsentry/relay/pull/4428))
+- Remove use of legacy project cache. ([#4419](https://github.com/getsentry/relay/pull/4419))
 
 ## 24.12.1